fix: fix coil and contact modal implemetation
