+++ src/renderer/components/_features/[workspace]/editor/graphical/ladder/index.tsx
@@ -3,8 +3,8 @@ import { Rung } from '@root/renderer/components/_organisms/rung'
 import { useState } from 'react'
 
 import BlockElement from '../elements/block'
-import { CoilElement } from '../elements/coil'
-import { ContactElement } from '../elements/contact'
+import CoilElement from '../elements/coil'
+import ContactElement from '../elements/contact'
 
 export default function LadderEditor() {
   const [rungs, setRungs] = useState<string[]>([])
